config/requirements.txt
psycopg2-binary==2.7.7
pytest==4.1.1
PyYAML==3.13
requests==2.21.0
SQLAlchemy==1.2.17
